Update ec2-pod roles to use explicit credentials
We cannot rely on Boto finding the credentials in the environment
because the environment is tucked away inside a Celery task. This means
both the standard ansible ec2 and ec2_* modules need to have
aws_access_key and aws_secret_key explicitly specified.  Additionally
the ec2.py filter plugin connects to AWS,  so make sure we're passing in
explicit references to the key/secret stored in the user's aws profile