Refactor ansible run_playbook and cumulus_log
This refactors ansible to use Popen so it is GPL compliant, refactor
cumulus_log to use environment variables because it no longer has access
to extra_vars which are passed in on the command line.

This also fixes cumulus_log to NOOP if a playbook is called from the
command line. The playbook still runs,  but no logging to the girder
cluster happens.