PUT back to complete attaching volume to cluster
This sends a PUT back to girder so girder can complete the process of
attaching a volume to a cluster.  This makes a two-way reference from
volume to cluster and from cluster to volumes.

We also add a block/rescue to attach.yml to post back an error status if
anything goes wrong.