Add detatch volume functionality
This follows the same kind of code path as attach. This means we call
detatch,  check with boto to make sure we're actually attached (so we
can fail early)  then launch ansible to detatch the volume.  Once we've
detached inside ansible we PUT back to detach/complete  to clean up the
girder models.

Its worth noting that this will fail if the directory is in use on the
remote machine,  when it fails there will be a message in
volume['ec2']['msg'] but it will be rather cryptic (ie.  failed,  please
check logs).  This is not ideal behavior.