Create cluster_adapter status API
This implements a _state_machine attribute on cluster adapters that is
used by cluster_adapter.status properties (getter/setters) to enforce
valid transitions from one cluster state to another.  _state_machine is
implemented as a ClusterStatus object and uses the cluster_adapter's
cluster['status'] key as its backing store.

This can be a little confusing. Cluster adapters have a _state_machine
attribute. The _state_machine doesn't maintain the state of an actual
cluster,  the AbstractClusterAdapter.cluster['status'] attribute does
this. The AbstractClusterAdapater delegates the responsibility of
transitioning the state to a ClusterStatus object (stored in
_state_machine). All this is hidden by AbstractClusterAdapter's status
property.