Made progress on volume test failures.
Added missing mock calls for create_volume and replaced the self._cumulus
user id with the self._user user id. This either changes the test case
failure to the actual difference in an assert statement or in a few cases
a ValueError('Required parameter id not set',) coming from the AWS api.