Update integration tests to handle Rackspace profiles
This add's a '-t' flag that can be either 'aws' or 'rax' (defaults to
'aws).

Basically we just refactor the Proxy() object into AWSProxy and
RAXProxy objects which are generated based on the -t flag. These just
overwrite methods that convert a config section into a POST-able
dictionary that is consistent with the correct profile schema.