Implement Rackspace cluster creation testing
Implements get_cluster_body() on RAXProxy inside the integration
framework.  This also introduces a [default] section with a single
key/value pair  'type' which can be one of 'aws' or 'rax.'

By default this is 'aws.' However,  if specified in the configuration
you do not have to pass the '-t' or '--type' flag to specify that you
are working with a RAXProxy rather than a AWSProxy object.

Finally this also ignores  rax.cfg in the tests/integration directory
allowing you to create a custom rax script for testing.