Add a code path for creating Rackspace clusters
This does not implement logic for launching clusters,  it just adds
the 'rax' ClusterType and code for creating a rax cluster model. This
follows the same format as an ec2 cluster (because both will use
ansible). Eventually we will want to try and merge these two code
paths for creating clusters (as they contain a great deal of duplicate
code and are using the same underlying system - ansible).