Explicitly import the cumulus plugin models
The models are no longer registered automatically as of girder
3.x, so they must be imported explicitly.

Note: if the plugin is not available at the time of loading
the script, we may need to move the import statements into
the functions instead of placing them at the top of the file.

It is also possible for us to import the models and register
them in a cumulus __init__ script, and then use model importer
like we used to.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>