Import required models in the functions
Specifically, in cumulus/common/girder.py, the AwsModel
would fail to be imported when the server would start.
This is because some files that import this one do not
have access to the AwsModel within the cumulus_plugin.

The only function that needs the AwsModel, `_get_profile()`,
is always called from within the cumulus_plugin. So it can
be assumed that the model will be registered and can be
accessed via ModelImporter.

I also moved the model import statements in
tasfklow/cluster/__init__.py to be within the function
that needed them to prevent possible import errors.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>