Add TokenScopes to access decorators
If a girder api key is used in which "Allow all actions on behalf of my
user" is not selected, token scopes are required in order for
`getCurrentUser()` to not return `None`. When the token scopes are not
present, various errors at different points where `getCurrentUser()` is
called/used occur.

Add the token scopes so that these errors do not occur.

Additionally, the `@access.token` decorators were changed to
`@access.user` decorators, since `@access.user` is usually what is
needed. `@access.token` is primarily used for tokens that may or may not
be associated with a user. But in nearly every case, we do want the
tokens to be associated with users, so `@access.user` is preferable.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>